requirements.txt
streamlit
openai-whisper
mistralai<2.0.0